# nexuz-ui-develop-p1-w1/packages/sdk/swagger.yaml
openapi: 3.0.3
info:
  title: NEXUZ API
  description: >-
    API for retrieving bundle information based on flight details and passenger
    information. It also retrieves seatMap based on the cabin requested by the
    passenger for its respective flights and provides details of available
    ancillaries for a particular flight.
  version: 1.0.0
...
paths:
...
  /search/calendar-fares:
    get:
      tags:
        - Search
      summary: Retrieve calendar fare date range
      description: >-
        Retrieves the lowest fare details for each date and cabin within a
        specified calendar date range.
      parameters:
        - in: query
          name: promotionCode
          description: Promotional code
          schema:
            type: string
            example: ABC123
          required: false
        - in: query
          name: routes
          description: |
            Comma-separated airport codes representing the route.
            - One-way / Round-trip: `"NRT,BKK"` (origin,destination)
            - Connecting: `"BKK,NRT,SIN"` (origin,stopover,destination)
          schema:
            type: string
            example: NRT,BKK
          required: true
        - in: query
          name: departureDateFrom
          description: Outbound departure date in `yyyy-MM-dd` format
          schema:
            type: string
            example: "2026-06-01"
          required: true
        - in: query
          name: departureDateTo
          description: |
            Inbound date in `yyyy-MM-dd` format for round-trips
          schema:
            type: string
            example: "2026-06-02"
          required: false
        - in: query
          name: language
          description: Language code
          schema:
            type: string
            example: en
          required: true
        - in: query
          name: currency
          description: Currency code
          schema:
            type: string
            example: JPY
          required: true
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-002-search_CalendarFaresResponse"
              examples:
                oneWay:
                  summary: One-way / One-way connecting scenario
                  value:
                    data:
                      outbound:
                        - cabin: ZIPFULLFLAT
                          dates:
                            - date: "2026-05-25"
                              baseFareForPromotion: 43535.01
                              lowestPrice: 50000.01
                        - cabin: STANDARD
                          dates:
                            - date: "2026-05-25"
                              baseFareForPromotion: 43535.01
                              lowestPrice: 50000.01
                roundTrip:
                  summary: Round-trip scenario
                  value:
                    data:
                      outbound:
                        - cabin: ZIPFULLFLAT
                          dates:
                            - date: "2026-05-25"
                              baseFareForPromotion: 43535.01
                              lowestPrice: 50000.01
                        - cabin: STANDARD
                          dates:
                            - date: "2026-05-25"
                              baseFareForPromotion: 43535.01
                              lowestPrice: 50000.01
                      inbound:
                        - cabin: ZIPFULLFLAT
                          dates:
                            - date: "2026-05-25"
                              baseFareForPromotion: 43535.01
                              lowestPrice: 50000.01
                        - cabin: STANDARD
                          dates:
                            - date: "2026-05-25"
                              baseFareForPromotion: 43535.01
                              lowestPrice: 50000.01
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-002-search_ErrorResponse"
              example:
                code: NEXUZCMNE001
                description: Required Input is Missing
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-002-search_ErrorResponse"
              examples:
                boundNotAvailable:
                  value:
                    code: NEXUZR002E051
                    description: Requested Bound is not available
                noValidData:
                  value:
                    code: NEXUZR002E052
                    description: No valid data found in the response
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-002-search_ErrorResponse"
              example:
                code: NEXUZR002E050
                description: Failed to RetrieveFareQuoteDateRange
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-002-search_ErrorResponse"
              examples:
                systemError:
                  value:
                    code: NEXUZCMNE002
                    description: System Error Occured
                internalError:
                  value:
                    code: NEXUZCMNE003
                    description: Internal Error Occured
...
  /offers/bundles:
    post:
      tags:
        - Offers
      summary: Retrieve bundles available for specified flight details
      description: >-
        Returns available service bundles (e.g., meals+baggage, etc.) for given
        flight segments and passenger composition.
      operationId: retrieveOfferBundles
      parameters:
        - name: currency
          in: query
          required: true
          schema:
            type: string
            example: JPY
      requestBody:
        required: true
        description: Request payload containing route, passenger, and flight details.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NEXUZ-R-004-offers_BundleRequest"
      responses:
        "200":
          description: Bundle retrieval successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-004-offers_BundleResponse"
        "400":
          description: Bad Request - Required Input is Missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-004-offers_ErrorResponse"
              examples:
                missingRequiredInput:
                  summary: Required input is missing
                  value:
                    code: NEXUZCMNE001
                    description: Required Input is Missing
        "404":
          description: Not found - Required Input is Missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-004-offers_ErrorResponse"
              examples:
                missingRequiredInput:
                  summary: no valid date is found
                  value:
                    code: NEXUZCMNE004
                    description: No Valid Data found
        "422":
          description: >-
            Unprocessable Entity - Request cannot be processed due to business
            validation issues.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-004-offers_ErrorResponse"
              example:
                code: NEXUZR004E002
                description: Failed to Retrieve service bundle details
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-004-offers_ErrorResponse"
              examples:
                internalError:
                  summary: Internal error
                  value:
                    code: NEXUZCMNE003
                    description: Internal Error Occurred
...
  /payment/{statusCheckKey}/status:
    post:
      tags:
        - Payment Status
      summary: Retrieve payment status
      description: >
        This endpoint is responsible for retrieving the payment status from POP
        and updating other systems based on the status, and returning a
        consolidated status to the UI.
      operationId: retrievePaymentStatus
      parameters:
        - name: statusCheckKey
          in: path
          required: true
          description: Represents order identifier to look up the payment order in POP.
          schema:
            type: string
        - name: currency
          in: query
          required: true
          description: Represents reservation currency.
          schema:
            type: string
          example: JPY
        - name: language
          in: query
          required: true
          description: Represents reservation language.
          schema:
            type: string
          example: ja
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NEXUZ-R-006-payment_PaymentStatusRequest"
      responses:
        "200":
          description: Consolidated response returned to UI.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-006-payment_PaymentStatusResponse"
              examples:
                success:
                  summary: Success status
                  value:
                    data:
                      status: Success
                      confirmationNumber: ABC123
                pending:
                  summary: Processing or Created status
                  value:
                    data:
                      status: Pending
                successWithWarning:
                  summary: Success with warning
                  value:
                    warning:
                      code: NEXUZR006W001
                      description: Unable to proceed further
                    data:
                      status: Success
                      confirmationNumber: ABC123
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-006-payment_ErrorResponse"
              examples:
                requiredInputMissing:
                  value:
                    code: NEXUZCMNE001
                    description: Required input is missing
        "422":
          description: Unprocessable Content
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/NEXUZ-R-006-payment_ErrorResponse"
                  - $ref: >-
                      #/components/schemas/NEXUZ-R-006-payment_PaymentStatusResponse
              examples:
                failedToGeneratePopToken:
                  value:
                    code: NEXUZR006E050
                    description: Unable to proceed further
                failedToRetrievePaymentStatus:
                  value:
                    code: NEXUZR006E051
                    description: Unable to proceed further
                paymentStatusIsFailed:
                  value:
                    data:
                      status: Failed
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NEXUZ-R-006-payment_ErrorResponse"
              examples:
                systemError:
                  value:
                    code: NEXUZCMNE002
                    description: System error occurred
                internalError:
                  value:
                    code: NEXUZCMNE003
                    description: Internal error occurred
...
components:
  schemas:
...
    NEXUZ-R-002-search_CalendarFaresResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/NEXUZ-R-002-search_CalendarFaresData"
    NEXUZ-R-002-search_CalendarFaresData:
      type: object
      required:
        - outbound
      properties:
        outbound:
          $ref: "#/components/schemas/NEXUZ-R-002-search_BoundFareData"
        inbound:
          $ref: "#/components/schemas/NEXUZ-R-002-search_BoundFareData"
    NEXUZ-R-002-search_BoundFareData:
      type: array
      items:
        type: object
        required:
          - cabin
          - dates
        properties:
          cabin:
            type: string
            example: ZIPFULLFLAT
            description: Cabin type (e.g., STANDARD or ZIPFULLFLAT).
          dates:
            type: array
            description: Fare details for the bound grouped by travel date.
            items:
              $ref: "#/components/schemas/NEXUZ-R-002-search_FareDateEntry"
    NEXUZ-R-002-search_FareDateEntry:
      type: object
      required:
        - date
        - lowestPrice
      properties:
        date:
          type: string
          example: "2026-05-25"
          description: Travel date to be shown in the calendar.
        baseFareForPromotion:
          type: number
          format: double
          example: 43535.01
          description: Applicable when a promoCode is provided.
        lowestPrice:
          type: number
          format: double
          example: 50000.01
          description: Lowest fare amount for the cabin on the given date.
    NEXUZ-R-002-search_ErrorResponse:
      type: object
      required:
        - code
        - description
      properties:
        code:
          type: string
          description: Custom error code
        description:
          type: string
          description: Error description
...
    NEXUZ-R-004-offers_ErrorResponse:
      type: object
      description: Standard error response body returned when the API fails.
      required:
        - code
        - description
      properties:
        code:
          type: string
          description: Application-specific error code.
          example: NEXUZCMNE001
        description:
          type: string
          description: Human-readable explanation of the error.
          example: Required Input is Missing
    NEXUZ-R-004-offers_BundleRequest:
      type: object
      description: >-
        Request payload containing flight route, passenger composition, and
        flight segment details.
      required:
        - routes
        - adult
        - outbound
      properties:
        routes:
          type: string
          description: Comma-separated airport codes (e.g., BKK,NRT).
          pattern: ^[A-Z]{3},[A-Z]{3}(,[A-Z]{3})?$
          example: BKK,NRT
        adult:
          description: number of adult passengers.
          type: integer
          minimum: 1
          maximum: 9
          example: 1
        childA:
          description: number of child passengers in category A.
          type: integer
          example: 1
          minimum: 1
          maximum: 9
        childB:
          description: number of child passengers in category B.
          type: integer
          example: 1
          minimum: 1
          maximum: 9
        childC:
          description: number of child passengers in category C.
          type: integer
          example: 1
          minimum: 1
          maximum: 9
        infant:
          description: number of infant passengers.
          type: integer
          example: 1
          minimum: 1
          maximum: 9
        outbound:
          type: array
          description: >
            Flight details for the outbound journey. For one-way trips, only
            `outbound` is required. For round-trips, both `outbound` and
            `inbound` are required.
          minItems: 1
          items:
            $ref: "#/components/schemas/NEXUZ-R-004-offers_Flight"
        inbound:
          type: array
          description: >
            Flight details for the inbound journey. Required only for
            round-trips.
          minItems: 1
          items:
            $ref: "#/components/schemas/NEXUZ-R-004-offers_Flight"
    NEXUZ-R-004-offers_Flight:
      description: >-
        Flight details including departure date, cabin class, logical flight ID,
        fare basis code, and fare class.
      type: object
      required:
        - departureDate
        - cabin
        - lfid
      properties:
        departureDate:
          description: Departure date in `2026-06-28T23:10:00` format.
          type: string
          format: date
          example: "2026-06-28"
        cabin:
          description: >-
            Cabin class for the flight. Valid values are "STANDARD" and
            "ZIPFULLFLAT".
          type: string
          enum:
            - STANDARD
            - ZIPFULLFLAT
        lfid:
          description: Logical Flight ID (LFID) representing the flight segment.
          type: integer
          example: 12345
          minimum: 10000
          maximum: 99999
        fareBasisCode:
          description: Fare basis code for the flight.
          type: string
          example: FQYB45
        fareClass:
          description: Fare class for the flight.
          type: string
          example: U
    NEXUZ-R-004-offers_BundleResponse:
      type: object
      required:
        - data
      properties:
        data:
          description: Bundle information organized by flight segments.
          type: array
          items:
            $ref: "#/components/schemas/NEXUZ-R-004-offers_SegmentWiseBundle"
    NEXUZ-R-004-offers_SegmentWiseBundle:
      description: >-
        Bundle information for a specific flight segment, identified by its
        Logical Flight ID (LFID).
      type: object
      required:
        - bundles
        - lfid
      properties:
        lfid:
          description: Logical Flight ID (LFID) representing the flight segment.
          type: integer
          example: 74982
        bundles:
          description: List of bundles available for the flight segment.
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/NEXUZ-R-004-offers_Bundle"
    NEXUZ-R-004-offers_Bundle:
      description: >-
        Bundle information including bundle code and passenger types with their
        associated services.
      type: object
      required:
        - bundleCode
        - passengerTypes
      properties:
        bundleCode:
          description: Code representing the bundle.
          type: string
          example: VALU
        passengerTypes:
          description: >-
            List of passenger types and their associated services within the
            bundle.
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/NEXUZ-R-004-offers_PassengerType"
    NEXUZ-R-004-offers_PassengerType:
      description: >-
        Information about a specific passenger type within a bundle, including
        the type, quantity, and associated services.
      type: object
      required:
        - type
        - amount
        - actualQuantity
        - bundleQuantity
        - categoryId
        - serviceId
        - cutoffHours
        - categories
      properties:
        type:
          description: Type of the passenger.
          type: string
          example: adult
        amount:
          description: Total amount for the passenger type within the bundle.
          type: number
          format: double
          example: 25001
        bundleQuantity:
          description: Quantity of the passenger type within the bundle.
          type: integer
          example: 1
        actualQuantity:
          description: >-
            Actual quantity of the passenger type which is determined based on
            count of each service for the passenger type in the bundle.
          type: integer
          example: 10
        categoryId:
          description: id representing categoryId of bundle.
          type: integer
          example: 142
        serviceId:
          description: id representing service id of bundle.
          type: integer
          example: 3
        cutoffHours:
          description: id representing representing cutoffHours of bundle
          type: integer
          example: 290
        categories:
          description: >-
            List of service categories and their associated services for the
            passenger type within the bundle.
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/NEXUZ-R-004-offers_Category"
    NEXUZ-R-004-offers_Category:
      description: >-
        Information about a specific service category within a bundle, including
        the category name and associated services.
      type: object
      required:
        - category
        - services
      properties:
        category:
          description: Name of the service category such as "Baggage", "Seat", "Meal", etc.
          type: string
          example: meals
        services:
          description: List of services available within the category.
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/NEXUZ-R-004-offers_Service"
    NEXUZ-R-004-offers_Service:
      description: >-
        Information about a specific service within a category, including the
        service code, quantity available, and description.
      type: object
      required:
        - code
        - quantityAvailable
        - description
      properties:
        code:
          description: Code representing the service.
          type: string
          example: MVAL
        quantityAvailable:
          description: Quantity of the service available.
          type: integer
          example: 260
        description:
          description: Description of the service.
          type: string
          example: Vegetarian meal option
...
    NEXUZ-R-006-payment_PaymentStatusRequest:
      type: object
      description: Request body for retrieving payment status.
      required:
        - paymentReferenceId
      properties:
        paymentReferenceId:
          type: integer
          description: Reference ID for the payment transaction.
          example: 123
    NEXUZ-R-006-payment_PaymentStatusResponse:
      type: object
      description: Response body for retrieving payment status.
      required:
        - data
      properties:
        warning:
          $ref: "#/components/schemas/NEXUZ-R-006-payment_WarningItem"
        data:
          $ref: "#/components/schemas/NEXUZ-R-006-payment_PaymentOrder"
    NEXUZ-R-006-payment_WarningItem:
      type: object
      description: Standard warning returned when the API encounters partial failures.
      required:
        - code
        - description
      properties:
        code:
          type: string
          description: Custom warning code.
          example: NEXUZR006W001
        description:
          type: string
          description: Warning description.
          example: Unable to proceed further
    NEXUZ-R-006-payment_PaymentOrder:
      type: object
      description: Payment order representing the current state of the payment.
      required:
        - status
      properties:
        status:
          type: string
          description: The status of the payment.
          enum:
            - Success
            - Failed
            - Pending
        confirmationNumber:
          type: string
          description: >-
            PNR number associated with the payment. This field is present only
            when the status is 'Success'.
          example: ABC123
    NEXUZ-R-006-payment_ErrorResponse:
      type: object
      description: Standard error response body.
      required:
        - code
        - description
      properties:
        code:
          type: string
          description: Application-level error code.
          example: NEXUZCMNE001
        description:
          type: string
          description: Error description.
          example: Required input is missing
...